backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.14
uvicorn[standard]==0.24.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-decouple==3.8
